Refactor playground example, adding Profile in the process
