+++ examples/playground/src/profiling.rs
@@ -1,0 +1,12 @@
+struct Profiling {
+}
+
+impl Profiling {
+    pub fn start_frame() {
+
+    }
+
+    pub fn stop_frame() {
+
+    }
+}
