+++ src/light/geometry_sprite_normal_pass.rs
@@ -70,8 +70,8 @@ impl GeometrySpriteNormalPass {
     where
         E: Element,
     {
-        #[cfg(feature = "coarse-prof")]
-        coarse_prof::profile!("light::GeometrySpriteNormalPass::draw");
+        //#[cfg(feature = "coarse-prof")]
+        //coarse_prof::profile!("light::GeometrySpriteNormalPass::draw");
 
         let mut sprite_infos = self.sprite_infos.borrow_mut();
         let sprite_info = sprite_infos.get(texture)?;
