+++ src/pass/color_pass.rs
@@ -51,8 +51,8 @@ impl ColorPass {
     ) where
         E: Element,
     {
-        #[cfg(feature = "coarse-prof")]
-        coarse_prof::profile!("ColorPass::draw");
+        //#[cfg(feature = "coarse-prof")]
+        //coarse_prof::profile!("ColorPass::draw");
 
         gl::draw(&self.program, matrices, [], draw_unit, params);
     }
