+++ src/plot/pass.rs
@@ -24,8 +24,8 @@ impl PlotPass {
         font: &Font,
         batch: &mut PlotBatch,
     ) -> Result<(), FrameError> {
-        #[cfg(feature = "coarse-prof")]
-        coarse_prof::profile!("PlotPass::draw");
+        //#[cfg(feature = "coarse-prof")]
+        //coarse_prof::profile!("PlotPass::draw");
 
         let draw_params = DrawParams {
             blend: Some(Blend::default()),
